Split five more sync files by topic
  PropertyRentSyncSystem (984) -> state/lifecycle, capture, realize; and into
    its own folder, as every other multi-file system here has
  BuildSyncSystem/Realize (972) -> realize, match, owned sub-elements, sub-nets
  NetSyncSystem/MixedOperation (934) -> operation, match, build, curves
  GrowableSyncSystem/Realize (867) -> realize, match, guard
  NetSyncSystem/Intent (832) -> intent, command encoding

Pure code movement, verified member by member against the previous revision.
