+++ CS2MultiplayerMod/Game/Sync/Systems/Simulation/GrowableSyncSystem/Realize.cs
@@ -13,6 +13,12 @@ using CS2MultiplayerMod.Game.Sync.Infrastructure;
 
 namespace CS2MultiplayerMod.Game.Sync.Systems
 {
+    // Applying the host's zoned-building lifecycle: a building spawns, changes level, is removed,
+    // or has its condition and state corrected. A correction that arrives before the building does
+    // is held and retried rather than dropped.
+    //
+    // Finding the building a command refers to is in RealizeMatch.cs; keeping a client from
+    // growing its own is in RealizeGuard.cs.
     public partial class GrowableSyncSystem
     {
         /// <summary>Zone cells are 8 m; a lot's half-extent is therefore its cell count times four.</summary>
@@ -449,419 +455,6 @@ namespace CS2MultiplayerMod.Game.Sync.Systems
             if (has == wanted) return;
             if (wanted) EntityManager.AddComponent<T>(entity);
             else EntityManager.RemoveComponent<T>(entity);
-        }
-
-        /// <summary>
-        /// The building standing at an anchor. Positions are computed from the same road and block
-        /// geometry on both machines, so the tolerance only absorbs float noise and a terrain
-        /// height that was sampled independently.
-        /// </summary>
-        private Entity FindGrowableAt(float3 position, Entity prefab, long now)
-        {
-            var candidates = new NativeList<Entity>(16, Allocator.Temp);
-            try
-            {
-                _objectSearch.CollectNear(position, AnchorSearchRadius, candidates);
-
-                Entity best = Entity.Null;
-                float bestDistance = AnchorMatchDistance * AnchorMatchDistance;
-                bool bestIsExact = false;
-
-                for (int i = 0; i < candidates.Length; i++)
-                {
-                    Entity candidate = candidates[i];
-                    if (!IsLiveGrowable(candidate, now)) continue;
-
-                    float distance = math.distancesq(
-                        EntityManager.GetComponentData<global::Game.Objects.Transform>(candidate)
-                            .m_Position.xz, position.xz);
-                    if (distance > AnchorMatchDistance * AnchorMatchDistance) continue;
-
-                    // Prefer the named prefab, but stay tolerant of a different one: a building
-                    // that levelled up no longer carries the prefab a removal names, and that
-                    // removal still has to reach it.
-                    bool exact = prefab != Entity.Null &&
-                                 EntityManager.GetComponentData<PrefabRef>(candidate).m_Prefab == prefab;
-                    if (bestIsExact && !exact) continue;
-                    if (exact && !bestIsExact)
-                    {
-                        best = candidate;
-                        bestDistance = distance;
-                        bestIsExact = true;
-                        continue;
-                    }
-                    if (best != Entity.Null && distance > bestDistance) continue;
-                    best = candidate;
-                    bestDistance = distance;
-                }
-                return best;
-            }
-            finally
-            {
-                candidates.Dispose();
-            }
-        }
-
-        /// <summary>
-        /// Everything already standing on the lot this spawn wants. Compares the two lot rectangles
-        /// rather than the two pivots: buildings of different sizes conflict long before their
-        /// centres coincide, and two neighbours on one street share a centre-to-centre distance
-        /// that says nothing about whether they fit.
-        /// </summary>
-        private void CollectOverlapping(Entity prefab, float3 position, quaternion rotation,
-            NativeList<Entity> blockers)
-        {
-            blockers.Clear();
-            if (!EntityManager.HasComponent<BuildingData>(prefab)) return;
-            float2 wantedExtent = LotExtent(EntityManager.GetComponentData<BuildingData>(prefab).m_LotSize);
-            float reach = math.length(wantedExtent) + ZoneCellSize;
-
-            var candidates = new NativeList<Entity>(32, Allocator.Temp);
-            try
-            {
-                _objectSearch.CollectNear(position, reach, candidates);
-                for (int i = 0; i < candidates.Length; i++)
-                {
-                    Entity candidate = candidates[i];
-                    if (!EntityManager.Exists(candidate) ||
-                        !EntityManager.HasComponent<Building>(candidate) ||
-                        !EntityManager.HasComponent<global::Game.Objects.Transform>(candidate) ||
-                        !EntityManager.HasComponent<PrefabRef>(candidate) ||
-                        EntityManager.HasComponent<Temp>(candidate) ||
-                        EntityManager.HasComponent<Deleted>(candidate) ||
-                        EntityManager.HasComponent<Owner>(candidate)) continue;
-
-                    Entity candidatePrefab = EntityManager.GetComponentData<PrefabRef>(candidate).m_Prefab;
-                    if (!EntityManager.HasComponent<BuildingData>(candidatePrefab)) continue;
-
-                    global::Game.Objects.Transform transform =
-                        EntityManager.GetComponentData<global::Game.Objects.Transform>(candidate);
-                    float2 extent = LotExtent(
-                        EntityManager.GetComponentData<BuildingData>(candidatePrefab).m_LotSize);
-
-                    if (RectanglesOverlap(position, rotation, wantedExtent,
-                            transform.m_Position, transform.m_Rotation, extent))
-                        blockers.Add(candidate);
-                }
-            }
-            finally
-            {
-                candidates.Dispose();
-            }
-        }
-
-        /// <summary>
-        /// True when the host's building is the one already standing here. Same prefab on the same
-        /// lot is the redelivery case; a different prefab at the same anchor is a real level
-        /// difference and has to be resolved, not ignored.
-        /// </summary>
-        private bool AlreadySatisfied(NativeList<Entity> blockers, Entity prefab, float3 position,
-            long now)
-        {
-            for (int i = 0; i < blockers.Length; i++)
-            {
-                Entity blocker = blockers[i];
-                if (!IsAutonomousGrowable(blocker, now)) continue;
-                if (EntityManager.GetComponentData<PrefabRef>(blocker).m_Prefab != prefab) continue;
-                float distance = math.distancesq(
-                    EntityManager.GetComponentData<global::Game.Objects.Transform>(blocker)
-                        .m_Position.xz, position.xz);
-                if (distance <= AnchorMatchDistance * AnchorMatchDistance) return true;
-            }
-            return false;
-        }
-
-        /// <summary>A building a player placed, as opposed to one a simulation grew.</summary>
-        private Entity FirstPlayerPlaced(NativeList<Entity> blockers, long now)
-        {
-            for (int i = 0; i < blockers.Length; i++)
-            {
-                if (!IsAutonomousGrowable(blockers[i], now)) return blockers[i];
-            }
-            return Entity.Null;
-        }
-
-        private static float2 LotExtent(int2 lotSize) =>
-            new float2(lotSize.x, lotSize.y) * (ZoneCellSize * 0.5f) - OverlapTolerance;
-
-        /// <summary>
-        /// Separating-axis test between two rotated lot rectangles. Four axes suffice: the two
-        /// rectangles' own edge normals, which for rectangles are their local x and z.
-        /// </summary>
-        private static bool RectanglesOverlap(float3 centreA, quaternion rotationA, float2 extentA,
-            float3 centreB, quaternion rotationB, float2 extentB)
-        {
-            if (extentA.x <= 0f || extentA.y <= 0f || extentB.x <= 0f || extentB.y <= 0f) return false;
-
-            float2 rightA = math.normalizesafe(math.rotate(rotationA, new float3(1f, 0f, 0f)).xz,
-                new float2(1f, 0f));
-            float2 forwardA = math.normalizesafe(math.rotate(rotationA, new float3(0f, 0f, 1f)).xz,
-                new float2(0f, 1f));
-            float2 rightB = math.normalizesafe(math.rotate(rotationB, new float3(1f, 0f, 0f)).xz,
-                new float2(1f, 0f));
-            float2 forwardB = math.normalizesafe(math.rotate(rotationB, new float3(0f, 0f, 1f)).xz,
-                new float2(0f, 1f));
-
-            float2 delta = centreB.xz - centreA.xz;
-            return !(SeparatedOn(rightA, delta, rightA, forwardA, extentA, rightB, forwardB, extentB) ||
-                     SeparatedOn(forwardA, delta, rightA, forwardA, extentA, rightB, forwardB, extentB) ||
-                     SeparatedOn(rightB, delta, rightA, forwardA, extentA, rightB, forwardB, extentB) ||
-                     SeparatedOn(forwardB, delta, rightA, forwardA, extentA, rightB, forwardB, extentB));
-        }
-
-        private static bool SeparatedOn(float2 axis, float2 delta,
-            float2 rightA, float2 forwardA, float2 extentA,
-            float2 rightB, float2 forwardB, float2 extentB)
-        {
-            float reachA = math.abs(math.dot(rightA, axis)) * extentA.x +
-                           math.abs(math.dot(forwardA, axis)) * extentA.y;
-            float reachB = math.abs(math.dot(rightB, axis)) * extentB.x +
-                           math.abs(math.dot(forwardB, axis)) * extentB.y;
-            return math.abs(math.dot(delta, axis)) > reachA + reachB;
-        }
-
-        private bool IsLiveGrowable(Entity entity, long now) =>
-            EntityManager.Exists(entity) &&
-            EntityManager.HasComponent<Building>(entity) &&
-            EntityManager.HasComponent<PrefabRef>(entity) &&
-            EntityManager.HasComponent<global::Game.Objects.Transform>(entity) &&
-            !EntityManager.HasComponent<Temp>(entity) &&
-            !EntityManager.HasComponent<Deleted>(entity) &&
-            !EntityManager.HasComponent<Owner>(entity) &&
-            IsAutonomousGrowable(entity, now);
-
-        private string DescribeBlocker(Entity blocker, long now)
-        {
-            Entity prefab = EntityManager.GetComponentData<PrefabRef>(blocker).m_Prefab;
-            string name = PrefabIndexSafeName(prefab);
-            bool grown = IsAutonomousGrowable(blocker, now);
-            return (grown ? "a grown '" : "a placed '") + (name ?? "?") + "'";
-        }
-
-        private int SeedFor(GrowableLifecycleCommand command)
-        {
-            // The built entity keeps the low 16 bits as its PseudoRandomSeed, which is the variant.
-            // Zero is the one value the game's own random rejects, so it is nudged rather than
-            // passed through - a building with no seed at all would fail to pick a mesh.
-            int seed = command.RandomSeed;
-            return seed == 0 ? 1 : seed;
-        }
-
-        /// <summary>
-        /// Remembers that a building was just asked for at this spot. The definition does not
-        /// become an entity until a later phase, so the only way to recognise our own building when
-        /// it appears is the position we asked for it at.
-        /// </summary>
-        private void NoteSelfRealized(Entity prefab, float3 position,
-            GrowableLifecycleCommand command, long now)
-        {
-            if (_selfRealized.Count >= MaxSelfRealized) _selfRealized.RemoveAt(0);
-            _selfRealized.Add(new PendingRealizedSpawn
-            {
-                Prefab = prefab,
-                Position = position,
-                Expiry = now + SelfRealizedWindowMs,
-                Command = command,
-            });
-        }
-
-        private bool TryTakeSelfRealized(Entity prefab, float3 position, long now,
-            out GrowableLifecycleCommand command)
-        {
-            command = null;
-            for (int i = _selfRealized.Count - 1; i >= 0; i--)
-            {
-                PendingRealizedSpawn entry = _selfRealized[i];
-                if (entry.Expiry <= now) { _selfRealized.RemoveAt(i); continue; }
-                if (entry.Prefab != prefab ||
-                    math.distancesq(entry.Position.xz, position.xz) >
-                    AnchorMatchDistance * AnchorMatchDistance) continue;
-                command = entry.Command;
-                _selfRealized.RemoveAt(i);
-                return true;
-            }
-            return false;
-        }
-
-        /// <summary>
-        /// Removes zoned buildings this client grew by itself. Its spawner is held for as long as
-        /// the session is synchronized, so in normal running this finds nothing - but authority is
-        /// handed back whenever sync drops (a resync, a world reload), and anything grown in that
-        /// window would otherwise stand forever on a lot the host has its own plans for.
-        ///
-        /// Catching them as they appear is what keeps the invariant simple: on a client, every
-        /// zoned building came from the host.
-        /// </summary>
-        private void RejectLocallyGrownBuildings(long now)
-        {
-            if (_createdBuildings.IsEmptyIgnoreFilter) return;
-
-            NativeArray<Entity> entities = _createdBuildings.ToEntityArray(Allocator.Temp);
-            try
-            {
-                for (int i = 0; i < entities.Length; i++)
-                {
-                    Entity entity = entities[i];
-                    Entity prefab = EntityManager.GetComponentData<PrefabRef>(entity).m_Prefab;
-                    if (!IsAutonomousGrowable(entity, now)) continue;
-
-                    float3 position = EntityManager
-                        .GetComponentData<global::Game.Objects.Transform>(entity).m_Position;
-                    GrowableLifecycleCommand command;
-                    if (TryTakeSelfRealized(prefab, position, now, out command))
-                    {
-                        // The definition is now a real native building. This is the first point at
-                        // which its construction clock and state payload can be applied safely.
-                        ApplyConditionAndState(entity, command);
-                        EntityManager.AddComponent<Updated>(entity);
-                        if (_realizationValidations.Count >= MaxRealizationValidations)
-                        {
-                            SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                                .Create("growable realization validation overflow", "growable",
-                                    CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.StreamLoss)
-                                .About("realization validation queue")
-                                .Tried("nothing - the validation queue was full"));
-                        }
-                        else _realizationValidations.Add(new RealizationValidation
-                        {
-                            Building = entity,
-                            Prefab = prefab,
-                            Position = position,
-                            Expiry = now + RealizationValidationWindowMs,
-                        });
-                        continue;
-                    }
-
-                    EntityManager.AddComponent<Deleted>(entity);
-                    _rejectedLocal++;
-                    Mod.log.Warn("[MP] GrowableSync: this client grew '" +
-                                 PrefabIndexSafeName(prefab) + "' at " + Format(position) +
-                                 " on its own; removed (the host decides zoned buildings).");
-                    Diagnostics.FlightRecorder.Note("locally grown building rejected");
-                }
-            }
-            finally
-            {
-                entities.Dispose();
-            }
-        }
-
-        /// <summary>
-        /// A definition being accepted is not yet proof that the generated building joined the
-        /// road/service graph. Re-run native Updated initialization until the root has a road,
-        /// reciprocal road buffer, and the standard utility consumers. Persistent failure is
-        /// structural divergence and escalates to the existing world-repair path.
-        /// </summary>
-        /// <summary>When this pass last ran with the road pipeline able to deliver.</summary>
-        private long _lastValidationTickMs;
-
-        /// <summary>
-        /// A building is being validated for having joined its ROAD graph, and roads arrive through
-        /// the net pipeline. While that pipeline is held - terrain catching up, or a placement
-        /// waiting for a target - the road this building needs cannot arrive by definition, so the
-        /// window must not count down. It is the same fifteen seconds as the hold itself, so left
-        /// running it would ask for a world reload over a road the mod was still holding back.
-        /// </summary>
-        private void ExtendValidationWindowsWhileRoadsHeld(long now)
-        {
-            long heldMs = _lastValidationTickMs == 0 ? 0 : now - _lastValidationTickMs;
-            _lastValidationTickMs = now;
-            if (!NetworkDependenciesHeld || heldMs <= 0) return;
-            for (int i = 0; i < _realizationValidations.Count; i++)
-                _realizationValidations[i].Expiry += heldMs;
-        }
-
-        private void ValidateRealizedBuildings(long now)
-        {
-            ExtendValidationWindowsWhileRoadsHeld(now);
-            for (int i = _realizationValidations.Count - 1; i >= 0; i--)
-            {
-                RealizationValidation pending = _realizationValidations[i];
-                Entity building = pending.Building;
-                if (building == Entity.Null || !EntityManager.Exists(building) ||
-                    EntityManager.HasComponent<Deleted>(building))
-                {
-                    _realizationValidations.RemoveAt(i);
-                    continue;
-                }
-
-                bool connected = HasNativeRoadConnection(building, pending.Prefab);
-                bool utilities = HasExpectedUtilityConsumers(building, pending.Prefab);
-                if (connected && utilities)
-                {
-                    _realizationValidations.RemoveAt(i);
-                    continue;
-                }
-
-                if (pending.Expiry <= now)
-                {
-                    _realizationValidations.RemoveAt(i);
-                    Mod.log.Warn("[MP] GrowableSync: generated building '" +
-                                 PrefabIndexSafeName(pending.Prefab) + "' at " +
-                                 Format(pending.Position) + " did not join its road/service graph; " +
-                                 "requesting world repair.");
-                    Diagnostics.FlightRecorder.Note("growable realization invalid/resync");
-                    SyncInbox.RequestResync(CS2MultiplayerMod.Game.Diagnostics.ResyncReport
-                        .Create("growable building failed road/service realization", "growable",
-                            CS2MultiplayerMod.Game.Diagnostics.ResyncEvidence.MissingTarget)
-                        .About("building road/service graph")
-                        .Tried("re-ran native initialization for 15 s of attempts, not counting time roads were held back"));
-                    continue;
-                }
-
-                if (!EntityManager.HasComponent<Updated>(building))
-                    EntityManager.AddComponent<Updated>(building);
-                Building data = EntityManager.GetComponentData<Building>(building);
-                if (data.m_RoadEdge != Entity.Null && EntityManager.Exists(data.m_RoadEdge) &&
-                    !EntityManager.HasComponent<Updated>(data.m_RoadEdge))
-                    EntityManager.AddComponent<Updated>(data.m_RoadEdge);
-            }
-        }
-
-        private bool HasNativeRoadConnection(Entity building, Entity prefab)
-        {
-            if (!EntityManager.HasComponent<BuildingData>(prefab)) return true;
-            BuildingData data = EntityManager.GetComponentData<BuildingData>(prefab);
-            if ((data.m_Flags & global::Game.Prefabs.BuildingFlags.RequireRoad) == 0) return true;
-
-            Building live = EntityManager.GetComponentData<Building>(building);
-            Entity road = live.m_RoadEdge;
-            if (road == Entity.Null || !EntityManager.Exists(road) ||
-                !EntityManager.HasBuffer<ConnectedBuilding>(road)) return false;
-            DynamicBuffer<ConnectedBuilding> connected =
-                EntityManager.GetBuffer<ConnectedBuilding>(road, true);
-            for (int i = 0; i < connected.Length; i++)
-                if (connected[i].m_Building == building) return true;
-            return false;
-        }
-
-        private bool HasExpectedUtilityConsumers(Entity building, Entity prefab)
-        {
-            if (EntityManager.HasComponent<UnderConstruction>(building) ||
-                EntityManager.HasComponent<Abandoned>(building) ||
-                EntityManager.HasComponent<Destroyed>(building) ||
-                !EntityManager.HasComponent<ConsumptionData>(prefab)) return true;
-            return EntityManager.HasComponent<ElectricityConsumer>(building) &&
-                   EntityManager.HasComponent<WaterConsumer>(building);
-        }
-
-        /// <summary>
-        /// Drain rather than apply: applying would duplicate the building the host already has.
-        ///
-        /// The host's own send loops back through the local observers, so nearly everything drained
-        /// here is the host's own echo - routine, and the same thing every other sync system skips
-        /// on <c>OriginPlayerId</c>. Only a command another player authored is worth a warning; it
-        /// means a peer is authoring zoned buildings it has no authority over.
-        /// </summary>
-        private void SyncInboxDrop(int localPlayerId)
-        {
-            int foreign = 0;
-            SimulationCommandMessage message;
-            while (_incoming.TryDequeue(out message))
-                if (message.OriginPlayerId != localPlayerId) foreign++;
-            if (foreign == 0) return;
-            Mod.log.Warn("[MP] GrowableSync: host discarded " + foreign +
-                         " zoned-building command(s) from another player; only a host may author them.");
         }
     }
 }
